Do not send in an empty array of annotations on success (#29)
GitHub's API requires that all annotations have a start and end on line parameter. Currently, we are passing in an array of empty annotations, which causes the API to error. This corrects that issue by not merging in annotations to the update check payload unless the status is failure.

Fixes #28