only send GitHub offenses instead of entire json output (#41)
We were not actually sending GitHub our list of annotations, but rather the entire report generated from rubocop --format json. This caused there to be empty offenses in the annotations we were trying to create, which the API rejected due to missing keys.

This change ensures that we actually send our annotation list, instead of the entire report. This also negated the need for the changes in #29, which have also been removed.

Fixes #40