Bugfix/respect rubocop level (#72)
Pass the exit code of the Rubocop run in the report as a synthetic JSON entry. Check it and respect it.

This respects whichever behavior Rubocop chooses to do to fail the check on --fail-level. Rubocop's semantics do not use the report output for this, but instead set the process status code because they are usually part of a build process.

Fixes #63