Find modified files with respect to base commit of branch (#139)
The problem with git diff origin/master when detecting modified files is that more recent changes in master that the current branch is not aware of will also be included. The outcome is that RuboCop will end up warning about files that have been fixed in master but not touched in the current branch.

Using git diff-tree we can find only the files that changed in the current branch which feels more consistent with how GitHub produces the diff in a Pull Request.